# openapi.yaml
openapi: 3.1.0
info:
  title: ECHO Platform API
  version: 1.0.0
  description: |
    ECHO Platform REST API and WebSocket specification.
    
    **Architecture**: Client-server relay with E2E encryption. The API server is a stateless
    relay and operational cache. It cannot read message content (E2E encrypted blobs only),
    does not own user identities (holders use did:key), and does not control token balances
    (Constellation metagraph).
    
    **Authentication**: All endpoints except /v1/auth/* require a Bearer token in the
    Authorization header. Tokens are issued after passkey/biometric verification.
    
    **Encryption**: Request/response bodies for sensitive endpoints are encrypted with
    a session key (AES-256-GCM) negotiated during auth. Message content is always E2E
    encrypted on the client (X25519 + ChaCha20-Poly1305) before transmission.
    
    **On-chain references**: Endpoints that interact with blockchain state (tokens, identity,
    trust) return cached values with a `snapshotHash` field for client-side verification
    against on-chain state (Phase 3+).
  contact:
    name: ECHO Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.echo.network/v1
    description: Production
  - url: https://api.staging.echo.network/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /users/profile:
    get:
      tags: [Users]
      summary: Get own profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    put:
      tags: [Users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 64
                username:
                  type: string
                  maxLength: 32
                  pattern: "^[a-zA-Z0-9_]+$"
                bio:
                  type: string
                  maxLength: 256
                showLastSeen:
                  type: string
                  enum: [everyone, contacts, nobody]
                showOnlineStatus:
                  type: string
                  enum: [everyone, contacts, nobody]
                showReadReceipts:
                  type: boolean
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /data-l1/merkle-roots:
    post:
      tags: [Identity]
      summary: Submit a Merkle root to local Data L1 (testnet)
      description: |
        Proxies a `MerkleRootUpdate` to the Constellation Data L1 (`DATA_L1_URL`).
        Used by `scripts/validate-phase1.sh` (WO-230). Server-side pre-validation (WO-35)
        runs before the L1 submit when optional batch metadata is supplied.
        No auth on local testnet; do not expose publicly without network controls.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [root, leafCount]
              properties:
                root:
                  type: string
                  description: 64-character lowercase hex SHA-256 Merkle root
                  pattern: "^[0-9a-f]{64}$"
                leafCount:
                  type: integer
                  minimum: 1
                leaf_count:
                  type: integer
                  minimum: 1
                  description: Alternative to leafCount (accepted by the API for shell convenience)
                timeFrom:
                  type: string
                  format: date-time
                  description: RFC3339 batch window start (optional; requires `timeTo`)
                timeTo:
                  type: string
                  format: date-time
                  description: RFC3339 batch window end (optional; requires `timeFrom`)
                schemaVersion:
                  type: integer
                  minimum: 0
                  description: Client schema version; must not exceed server max (WO-35)
      responses:
        "201":
          description: Accepted and forwarded to Data L1
          content:
            application/json:
              schema:
                type: object
                properties:
                  tx_id:
                    type: string
                  txHash:
                    type: string
        "400":
          $ref: "#/components/schemas/Error"
        "502":
          description: Data L1 returned an error
        "503":
          description: DATA_L1_URL not configured

  # ──────────────────────── Phase 1 validation (local dev) ────────────────────────
  /phase1/trust-tier-commitment:
    post:
      tags: [Identity]
      summary: Anchor H(tier||nonce) on Identity L1 (WO-230 dev script only)
      description: |
        Forwards `TrustTierCommitmentUpdate` to Identity Metagraph L1. Enabled only when
        `ENVIRONMENT=development` or `PHASE1_ALLOW_OPEN_VALIDATE=true`. Do not expose in production.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject_did, tier, nonce]
              properties:
                subject_did:
                  type: string
                  description: Holder did:key (e.g. from Step 2)
                tier:
                  type: integer
                  minimum: 1
                  maximum: 5
                nonce:
                  type: string
                  description: Random nonce (commitment = SHA256(byte(tier) || nonce))
      responses:
        "200":
          description: Submitted to Identity L1
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject_did:
                    type: string
                  tier:
                    type: integer
                  commitment:
                    type: string
                  anchored_at:
                    type: integer
                    format: int64
                  tx_hash:
                    type: string
        "400":
          $ref: "#/components/schemas/Error"
        "403":
          description: Not allowed outside development
        "502":
          description: Identity L1 reject or transport error
        "503":
          description: Identity L1 client not configured

  # ──────────────────────── IDENTITY ────────────────────────
  /identity/register:
    post:
      tags: [Identity]
      summary: Register a did:key with the local Identity Service
      description: |
        Phase-1 (per ADR-0001) DID registration endpoint. The client derives the
        canonical `did:key:z…` from a P-256 public key (the
        `pkg/didkey.DeriveFromPublicKeyHex` helper, or any W3C-compliant
        did:key implementation) and posts both the DID and the underlying
        public key for binding.

        The handler re-derives the canonical DID from `public_key_hex` and
        rejects the request if it does not match the supplied `did`. This
        guarantees no caller can register a DID for a key they do not encode
        correctly. Possession of the *private* key is proven later, during
        passkey signing, and on every subsequent authenticated request.

        Idempotent: re-posting the same `(did, public_key_hex)` pair returns
        `200 OK` with `existing: true` instead of `409`.

        No auth required — the DID *is* the public key, so registration is
        a public bootstrap step (rate-limited).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityRegisterRequest"
      responses:
        "201":
          description: DID newly registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityRegisterResponse"
        "200":
          description: DID already registered with the same public key (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityRegisterResponse"
        "400":
          description: |
            Validation error. Possible `code` values:
            * `MISSING_DID` — `did` field empty
            * `MISSING_PUBLIC_KEY` — `public_key_hex` empty
            * `INVALID_PUBLIC_KEY` — `public_key_hex` is not a valid SEC1
              uncompressed P-256 point
            * `DID_KEY_MISMATCH` — `did` does not match the canonical
              derivation from `public_key_hex`
            * `MALFORMED_BODY` — request body is not valid JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: |
            DID is already registered with a *different* public key
            (`code: DID_ALREADY_REGISTERED`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Registry not configured (`code: REGISTRY_NOT_CONFIGURED`)"
...
  /identity/devices:
    post:
      tags: [Identity]
      summary: Register an additional device key (WO-273)
      description: |
        Adds a new P-256 device public key for an existing subject DID. The
        request body must be signed with ECDSA P-256 over SHA-256: the signer
        uses the private key for `signing_did`, which must already be a
        registered device for `subject_did`. Send the signature as lowercase hex
        in the `X-Identity-Signature` header (IEEE P1363 64-byte R||S or ASN.1
        DER).

        Constellation Identity Metagraph anchoring of device events is not yet
        wired; persistence is PostgreSQL (or in-memory registry) only.
      security: []
      parameters:
        - name: X-Identity-Signature
          in: header
          required: true
          description: Hex-encoded ECDSA signature over the raw JSON body bytes
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityAddDeviceRequest"
      responses:
        "200":
          description: Device added; returns updated device list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityResolveResponse"
        "400":
          description: Validation error (missing fields, malformed signature hex, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unknown signing device or invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: subject_did not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Device public key already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Registry not configured

  # ───── Legacy/Phase-0 DID anchor endpoints (retained for compatibility) ─────
...
  /identity/credentials:
    get:
      tags: [Identity]
      summary: List verifiable credentials
      description: Returns credentials for the authenticated subject (WO-273). Requires issuer stack and same env as `cmd/credentials`.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Credential list
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: "#/components/schemas/Credential"
        "503":
          description: Credential service not configured
    post:
      tags: [Identity]
      summary: Request a new credential
      description: |
        Issues a VC for the authenticated subject via `pkg/credentials` (WO-273).
        `subjectDid` defaults to the JWT `sub`; if set, it must equal `sub`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schemaId]
              properties:
                schemaId:
                  type: string
                  description: VC type / schema id (`ProofOfHumanity`, `KYCLite`, `HighAssurance`, `Professional`)
                credentialType:
                  type: string
                  description: Optional alias for schemaId
                subjectDid:
                  type: string
                  description: Optional; must match Bearer principal when set
                claims:
                  type: object
                  additionalProperties: true
                verificationClaims:
                  type: array
                  items:
                    type: object
                preferredFormat:
                  type: string
                  enum: [json-ld, jwt, sd-jwt]
                expirationYears:
                  type: integer
                  minimum: 1
      responses:
        "202":
          description: Credential issued (sync completion; 202 for OpenAPI contract)
        "400":
          $ref: "#/components/schemas/Error"
        "503":
          description: Credential service not configured (missing issuer keys, verifier DID, Identity L1 URL, or OIDC4VC base URLs when OIDC is enabled)
...
  /identity/trust-tier/commitment:
    post:
      tags: [Identity]
      summary: Anchor H(tier ‖ nonce) trust-tier commitment (WO-274)
      description: |
        Submits TrustTierCommitmentUpdate to Identity L1 and caches the result in Redis (~60s TTL)
        when Redis is configured. Requires Bearer token; subject_did must match the token principal.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier, nonce]
              properties:
                subject_did:
                  type: string
                  description: Optional; defaults to the authenticated DID
                tier:
                  type: integer
                  minimum: 1
                  maximum: 5
                nonce:
                  type: string
      responses:
        "200":
          description: Commitment accepted by L1
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject_did:
                    type: string
                  tier:
                    type: integer
                  commitment:
                    type: string
                    description: 64-char lowercase hex SHA-256 digest
                  anchored_at:
                    type: integer
                    format: int64
                    description: Unix time in milliseconds
        "502":
          description: L1 submit failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from /auth/verify
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        did:
          type: string
        displayName:
          type: string
        username:
          type: string
          nullable: true
        avatarURL:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          nullable: true
        trustTier:
          type: integer
          minimum: 1
          maximum: 5
        verificationBadges:
          type: array
          items:
            type: string
            enum: [email_verified, phone_verified, identity_verified, government_id, enterprise]
        isOnline:
          type: boolean
          nullable: true
          description: "null if privacy setting hides this"
        lastSeen:
          type: string
          format: date-time
          nullable: true
          description: "null if privacy setting hides this"
    Error:
      type: object
      description: |
        Canonical error envelope returned by all Phase-1 endpoints. The
        backend's WriteError helper guarantees this shape on any non-2xx
        response.
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. `INVALID_PUBLIC_KEY`).
          example: "INVALID_PUBLIC_KEY"
        message:
          type: string
          description: Human-readable diagnostic; safe to surface to end users.
          example: "did_key_mismatch: supplied did does not match canonical derivation from public_key_hex"
        request_id:
          type: string
          description: Echo of the X-Request-ID header for log correlation.
    IdentityRegisterRequest:
      type: object
      required: [did, public_key_hex]
      properties:
        did:
          type: string
          description: W3C did:key identifier the client claims for this binding.
          example: "did:key:z2dmzD81cgPx8Vki7JbuuMmFYrWPgYoytykUZ3eyqht1j9KbrL2g"
        public_key_hex:
          type: string
          description: |
            SEC1 uncompressed P-256 public key (65 bytes, `04` prefix) as
            lowercase hex. Must derive to the supplied `did` via the W3C
            did:key method (multicodec `p256-pub` 0x1200, multibase
            base58btc); mismatches are rejected with 400 DID_KEY_MISMATCH.
          pattern: "^04[0-9a-f]{128}$"
          example: "04abcd…"
    IdentityRegisterResponse:
      type: object
      required: [did, public_key_hex, registered_at, existing]
      properties:
        did:
          type: string
          example: "did:key:z2dmzD81cgPx8Vki7JbuuMmFYrWPgYoytykUZ3eyqht1j9KbrL2g"
        public_key_hex:
          type: string
          pattern: "^04[0-9a-f]{128}$"
        registered_at:
          type: string
          format: date-time
          description: RFC3339 nano timestamp of the *first* time this DID was registered.
        existing:
          type: boolean
          description: |
            `true` when the (did, public_key_hex) pair was already present and
            the server returned 200. `false` on a fresh 201.
...
    IdentityDeviceRecord:
      type: object
      required: [public_key_hex, device_label, registered_at]
      properties:
        public_key_hex:
          type: string
          pattern: "^04[0-9a-f]{128}$"
        device_label:
          type: string
          example: "primary"
        registered_at:
          type: string
          format: date-time
    IdentityResolveResponse:
      type: object
      required: [did, devices]
      properties:
        did:
          type: string
        devices:
          type: array
          items:
            $ref: "#/components/schemas/IdentityDeviceRecord"
    IdentityAddDeviceRequest:
      type: object
      required: [subject_did, new_public_key_hex, signing_did]
      properties:
        subject_did:
          type: string
        new_public_key_hex:
          type: string
          pattern: "^04[0-9a-f]{128}$"
        device_label:
          type: string
          description: Optional label; defaults to `device` when omitted.
        signing_did:
          type: string
          description: |
            did:key of an already-registered device; the request body is
            signed with that device's private key.
...
    Credential:
      type: object
      properties:
        id:
          type: string
        schemaId:
          type: string
        issuerDID:
          type: string
        status:
          type: string
          enum: [active, revoked, expired]
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        l1TxHash:
          type: string
          description: Optional Identity Metagraph L1 anchor reference for this credential
...